+++ packages/types/src/product.ts
@@ -1,0 +1,10 @@
+interface Product {
+    id: number;
+    title: string;
+    description: string;
+    gpsr: {
+        manufacturerName: string;
+        manufacturerAddress: string;
+        safetyContact: string;
+    };
+}